updated plugin name fetch usage to fix NPE when offline
